+++ src/main/java/com/revature/dao/ArtistDao.java
@@ -1,0 +1,6 @@
+package com.revature.dao;
+
+
+public interface ArtistDao
+{
+}
