+++ src/main/java/com/revature/dao/BandMemberDao.java
@@ -1,0 +1,6 @@
+package com.revature.dao;
+
+
+public interface BandMemberDao
+{
+}
