+++ src/main/java/com/revature/dao/UserDao.java
@@ -1,0 +1,12 @@
+package com.revature.dao;
+
+
+import com.revature.domain.User;
+
+public interface UserDao
+{
+    int createUser(String username, String password, int accountType);
+
+    User getUserById(int id);
+
+}
