+++ src/main/java/com/revature/dao/VenueOwnerDao.java
@@ -1,4 +1,4 @@
-package com.revature.doa;
+package com.revature.dao;
 
 
 public interface VenueOwnerDao
