--- src/main/java/com/revature/doa/UserDao.java
@@ -1,6 +1,0 @@
-package com.revature.doa;
-
-
-public interface UserDao
-{
-}
